+++ org-trello-query.el
@@ -24,7 +24,7 @@
 (defun* orgtrello-query/--standard-error-callback (&key response error-thrown &allow-other-keys)
   "Standard error callback which expects a RESPONSE.
 Simply displays an error message in the minibuffer with the error code."
-  (orgtrello-log/msg *OT/INFO* "client - Problem during the request to the proxy- error-thrown: %s" error-thrown)
+  (orgtrello-log/msg *OT/INFO* "client - Problem during request - error-thrown: %s" error-thrown)
   (orgtrello-log/msg *OT/DEBUG* "Detailed response: %S" response))
 
 (defun* orgtrello-query/--standard-success-callback (&key response &allow-other-keys)